+++ cliquet/tests/test_views_hello.py
@@ -30,7 +30,7 @@ class HelloViewTest(BaseWebTest, unittest.TestCase):
         eos = '2069-02-21'
         with mock.patch.dict(
                 self.app.app.registry.settings,
-                [('cliquet.eos', eos)]):
+                [('eos', eos)]):
             response = self.app.get('/')
             self.assertEqual(response.json['eos'], eos)